Automate the full Ghidra pipeline and refresh the corpus

+++ decomp/ST.exe/functions/00627390/decomp.c
@@ -8,8 +8,8 @@ void __fastcall FUN_00627390(STMineSetC *param_1)
   if ((param_1->field_02AE == CASE_0) || (param_1->field_02AE == CASE_1)) {
     param_1->field_02AE = CASE_2;
     param_1->field_030F =
-         ((int)(*(int *)(&DAT_007d02a4 + (uint)(byte)param_1->field_02AD * 4) +
-               (*(int *)(&DAT_007d02a4 + (uint)(byte)param_1->field_02AD * 4) >> 0x1f & 3U)) >> 2) +
+         ((int)(*(int *)(&DAT_007d02a4 + (uint)param_1->field_02AD * 4) +
+               (*(int *)(&DAT_007d02a4 + (uint)param_1->field_02AD * 4) >> 0x1f & 3U)) >> 2) +
          g_playSystem_00802A38->field_00E4;
   }
   return;
